refactors packages structure in schedule
