+++ src/main/java/com/intive/patronage/toz/schedule/ScheduleController.java
@@ -3,6 +3,11 @@ package com.intive.patronage.toz.schedule;
 import com.intive.patronage.toz.error.ArgumentErrorResponse;
 import com.intive.patronage.toz.error.ErrorResponse;
 import com.intive.patronage.toz.error.ValidationErrorResponse;
+import com.intive.patronage.toz.schedule.model.db.Reservation;
+import com.intive.patronage.toz.schedule.model.view.DayConfigView;
+import com.intive.patronage.toz.schedule.model.view.ReservationView;
+import com.intive.patronage.toz.schedule.model.view.ScheduleView;
+import com.intive.patronage.toz.schedule.service.ReservationService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
