+++ src/main/java/com/intive/patronage/toz/schedule/model/view/DayConfigView.java
@@ -1,4 +1,4 @@
-package com.intive.patronage.toz.schedule;
+package com.intive.patronage.toz.schedule.model.view;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.annotations.ApiModel;
@@ -11,7 +11,7 @@ import java.util.List;
 @ApiModel("Day configuration")
 @Getter
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-class DayConfigView {
+public class DayConfigView {
 
     @ApiModelProperty(value = "Name of day")
     private final DayOfWeek dayOfWeek;
