+++ src/main/java/com/intive/patronage/toz/schedule/model/view/ScheduleView.java
@@ -1,4 +1,4 @@
-package com.intive.patronage.toz.schedule;
+package com.intive.patronage.toz.schedule.model.view;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.annotations.ApiModel;
@@ -9,11 +9,11 @@ import java.util.List;
 @ApiModel("Schedule")
 @Getter
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-class ScheduleView {
+public class ScheduleView {
     private final List<ReservationView> reservations;
     private final List<DayConfigView> configs;
 
-    ScheduleView(List<ReservationView> reservations, List<DayConfigView> configs) {
+    public ScheduleView(List<ReservationView> reservations, List<DayConfigView> configs) {
         this.reservations = reservations;
         this.configs = configs;
     }
